requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
schedule==1.2.0
APScheduler==3.10.4

# AI & ML
groq==0.3.1
requests==2.31.0
openai==1.3.0  # For free alternative - we'll use Groq primarily
huggingface-hub==0.19.4
transformers==4.35.2

# Video Processing
moviepy==1.0.3
Pillow==10.1.0
opencv-python==4.8.1.78

# Audio Processing
gtts==2.4.0  # Google Text-to-Speech (free)
pydub==0.25.1

# YouTube API
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Data Sources
praw==7.7.1  # Reddit API
pyyoutube==2.3.2  # YouTube Data API alternative
pytrends==5.0.3  # Google Trends

# Email & Reporting
smtplib  # Built-in
email  # Built-in

# Database (for tracking videos, stats)
sqlite3  # Built-in

# Utilities
python-dateutil==2.8.2
pytz==2023.3
youtube-transcript-api==0.6.1

